Ended obligatory ? part

+++ srcs/formats/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: lperson- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 10:45:05 by lperson-          #+#    #+#             */
-/*   Updated: 2019/11/05 11:55:00 by lperson-         ###   ########.fr       */
+/*   Updated: 2019/11/06 03:35:13 by lperson-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 #include "buffer.h"
 #include "lft_string.h"
 
-size_t			count_digits(int nbr, char *base)
+size_t			count_digits(long nbr, char *base)
 {
 	size_t			count;
-	unsigned int	nb;
+	unsigned long	nb;
 	size_t			len;
 
 	count = 0;
@@ -33,7 +33,7 @@ size_t			count_digits(int nbr, char *base)
 	return (count);
 }
 
-int				ft_putnbr_base(char *buffer, unsigned int nbr, char *base)
+int				ft_putnbr_base(char *buffer, unsigned long nbr, char *base)
 {
 	int				bytes;
 	size_t			len;
